@msokolov 2020-12-30
At @rmuir's suggestion, I re-coded as JMH benchmark, with a similar result:

```
Benchmark                                           Mode  Cnt     Score    Error  Units
TestIndexInputReadFloats.testIndexInputReadFloats  thrpt   25  1371.856 ± 11.681  ops/s
TestLuceneBaseline.testLuceneBaseline              thrpt   25   572.370 ±  3.918  ops/s
TestOnHeapArray.testOnHeapArray                    thrpt   25  1915.688 ±  1.233  ops/s
```

It seems as if a large benefit can be had by implementing `IndexInput.readFloats`, so I'll work that up, but there's still some fairly big gap with on-heap arrays. Maybe some kind of loop-unrolling / vectorization? I'll look at asm and see what I can tease out of that.